GWAF-package {GWAF}	R Documentation

Genome-Wide Association analyses with Family data

Description

Functions to fit Linear Mixed Effects (LME) model accounting for within pedigree correlation for testing genetic association between genotyped/imputed SNPs and a continuous trait and functions to fit Generalized Estimation Equation (GEE) model using each pedigree as a cluster for testing genetic association between SNPs and a dichotomous trait. In addition, GWAF package also provides functions for making genome-wide p-values plot and QQ plot containing genomic control parameter estimate.

Details

Package: 	 GWAF

Type: 	 Package

Version: 	 2.0

Date: 	 2009-09-22

License: 	 GPL (>= 2)

Author(s)

Qiong Yang <qyang@bu.edu> and Ming-Huei Chen <mhchen@bu.edu>

Maintainer: Ming-Huei Chen <mhchen@bu.edu>

[Package GWAF version 1.1 Index]