auto {GWAF}	R Documentation

function to generate scripts for genome-wide association analysis using GWAF

Description

Given a phenotype file, a pedigree file, a phenotype of interest, covariates, analysis of interest (can be 'lme', 'lme.imputed', 'gee' or 'gee.imputed'), a path or directory (genopath) that keeps geotype files, and other arguments, auto function generates one R script and one shell script that excutes the R script for each genotype file, and one list file that excutes all shell scripts in batch mode.

Usage

auto(genopath, phenfile, pedfile, outfile, phen, covars = NULL, analysis, lib.loc, model = NULL, 
kinmat = NULL, col.names = F, sep.ped = ",", sep.phe = ",", sep.gen = ",")

Arguments

genopath 	 a character string indicating the path or directory that keeps genotype files to be analyzed

phenfile 	 a character string naming the phenotype file for reading (see format requirement in details)

pedfile 	 a character string naming the pedigree file for reading (see format requirement in details)

outfile 	 a character string naming the result file for writing

phen 	 a character string for a phenotype name in phenfile

covars 	 a character vector for covariates in phenfile

analysis 	 a character string indicating the analysis of interest available in GWAF package, can be 'lme', 'lme.imputed', 'gee' or 'gee.imputed'

lib.loc 	 a character string indicating the location of GWAF package

model 	 a single character of 'a','d','g', or 'r', with 'a'=additive, 'd'=dominant, 'g'=general and 'r'=recessive models; Not appropriate/needed for analyzing imputed SNPs

kinmat 	 a character string naming the file where kinship coefficient matrix is kept; needed for LME analyses

col.names 	 a logical value indicating whether the output file should contain column names

sep.ped 	 the field separator character for pedigree file

sep.phe 	 the field separator character for phenotype file

sep.gen 	 the field separator character for genotype file

Details

auto function generates one R script and one shell script that excutes the R script for each genotype file, and one list file that can excute all shell scripts in batch mode to analyze all genotype data in genopath. These scripts are named based on the phenotype of interest, the analysis of interest and the time these scripts are generated. After creating these scripts, auto function gives a message telling the user how to submit ALL the jobs (using ksh XXXX.lst). When a submitted job is completed, a log file indicating which genotype file was analyzed will be generated and the R script and the shell script will be removed. The number of log files should equal to the number of genotype files in genopath, if all jobs are completed. All the results will be written and appended to the user specified single output file. Different outfile should be assigned for different genopath to avoid over-writting.

Value

No value is returned. Instead, results are written to outfile.

Author(s)

Ming-Huei Chen <mhchen@bu.edu> and Qiong Yang <qyang@bu.edu>

[Package GWAF version 1.2 Index]