gee.lgst {GWAF}	R Documentation

function for testing association between a dichotomous trait and a genotyped SNP in family data using GEE

Description

Fit logistic regression via GEE to test association between a dichotomous phenotype and one genotyped SNP in a genotype file with user specified genetic model. Each family is treated as a cluster, with independence working correlation matrix used in the robust variance estimator. This function is called in gee.lgst.batch function to apply association test to all SNPs in the genotype data.

Usage

gee.lgst(snp, phen, test.dat, covar=NULL, model="a")

Arguments

snp 	 genotype data of a SNP

phen 	 a character string for a phenotype name in phenfile

test.dat 	 the product of merging phenotype, genotype and pedigree data, should be ordered by "famid"

covar 	 a character vector for covariates in phenfile

model 	 a single character of 'a','d','g', or 'r', with 'a'=additive, 'd'=dominant, 'g'=general and 'r'=recessive models

Details

The gee.lgst function tests association between a dichtomous trait and a SNP from a dataset that contains phenotype, genotype and pedigree data, where the dataset needs to be ordered by famid.

Value

Please see output in gee.lgst.batch.

Author(s)

Qiong Yang <qyang@bu.edu> and Ming-Huei Chen <mhchen@bu.edu>

References

Liang, K.Y. and Zeger, S.L. (1986) Longitudinal data analysis using generalized linear models. Biometrika, 73 13–22.

Zeger, S.L. and Liang, K.Y. (1986) Longitudinal data analysis for discrete and continuous outcomes. Biometrics, 42 121–130.

Vincent J Carey.Ported to R by Thomas Lumley (versions 3.13 and 4.4) and Brian Ripley (version 4.13). gee: Generalized Estimation Equation solver. [4.13]. 2007. Ref Type: Computer Program, http://cran.r-project.org/

See Also

gee() function from package gee

[Package GWAF version 1.2 Index]