gee.lgst.batch.imputed {GWAF}	R Documentation

function to test genetic association between a dichotomous trait and a batch of imputed SNPs in families using GEE

Description

Fit logistic regression via GEE to test association between a dichotomous phenotype and all imputed SNPs in a genotype file. Each family is treated as a cluster, with independence working correlation matrix used in the robust variance estimator. This function applies the same trait-SNP association test to all SNPs in the imputed genotype data. The trait-SNP association test is carried out by gee.lgst.imputed function where the the gee() function from package gee is used.

Usage

gee.lgst.batch.imputed(genfile, phenfile, pedfile, outfile, phen, 
covars=NULL, col.names=T, sep.ped=",", sep.phe=",", sep.gen=",")

Arguments

genfile 	 a character string naming the genotype file for reading (see format requirement in details)

phenfile 	 a character string naming the phenotype file for reading (see format requirement in details)

pedfile 	 a character string naming the pedigree file for reading (see format requirement in details)

outfile 	 a character string naming the result file for writing

phen 	 a character string for a phenotype name in phenfile

covars 	 a character vector for covariates in phenfile

col.names 	 a logical value indicating whether the output file should contain column names

sep.ped 	 the field separator character for pedigree file

sep.phe 	 the field separator character for phenotype file

sep.gen 	 the field separator character for genotype file

Details

Similar to the details for gee.lgst.batch but here the SNP data contains imputed genotypes (allele dosages) that are continuous and range from 0 to 2. In addition, the user specified genetic model argument is not available.

Value

No value is returned. Instead, results are written to outfile.

phen 	 phenotype name

snp 	 SNP name

N 	 the number of individuals in analysis

Nd 	 the number of individuals in affected sample in analysis

AF 	 imputed allele frequency of coded allele

AFd 	 imputed allele frequency of coded allele in affected sample

beta 	 regression coefficient of SNP covariate

se 	 standard error of beta

remark 	 warning or additional information for the analysis, note that the genotype counts are based on rounded imputed genotypes; 'not converged' indicates the GEE analysis did not converge; 'logistic reg' indicates GEE model is replaced by logistic regression; 'exp count<5' indicates any expected count is less than 5 in phenotype-genotype table; 'not converged and exp count<5', 'logistic reg & exp count<5' are noted similarly; 'collinearity' indicates collinearity exists between SNP and some covariates

pval 	 p-value of the chi-square statistic

Author(s)

Qiong Yang <qyang@bu.edu> and Ming-Huei Chen <mhchen@bu.edu>

[Package GWAF version 1.2 Index]