lme.batch.imputed {GWAF}	R Documentation

function to test associations between a continuous trait and a batch of imputed SNPs in families using Linear Mixed Effects model

Description

Fit linear mixed effects model to test association between a continuous phenotype and all imputed SNPs in a genotype file. The SNP genotype is treated as fixed effects, and a random effect correlated according to degree of relatedness within a family is also fitted. In each trait-SNP assocaition test, the lmekin() function which is modified from the same named function in package kinship is used.

Usage

lme.batch.imputed(phenfile, genfile, pedfile, phen, kinmat, covars=NULL, 
outfile, col.names=T, sep.ped=",", sep.phe=",", sep.gen=",")

Arguments

phenfile 	 a character string naming the phenotype file for reading (see format requirement in details)

genfile 	 a character string naming the genotype file for reading (see format requirement in details)

pedfile 	 a character string naming the pedigree file for reading (see format requirement in details)

phen 	 a character string for a phenotype name in phenfile

kinmat 	 a character string naming the file where kinship coefficient matrix is kept

covars 	 a character vector for covariates in phenfile

outfile 	 a character string naming the result file for writing

col.names 	 a logical value indicating whether the output file should contain column names

sep.ped 	 the field separator character for pedigree file

sep.phe 	 the field separator character for phenotype file

sep.gen 	 the field separator character for genotype file

Details

Similar to the details for lme.batch function but here the SNP data contains imputed genotypes (allele dosages) that are continuous and range from 0 to 2. In addition, the user specified genetic model argument is not available.

Value

No value is returned. Instead, results are written to outfile.

phen 	 phenotype name

snp 	 SNP name

N 	 the number of individuals in analysis

AF 	 imputed allele frequency of coded allele

h2q 	 the portion of phenotypic variation explained by the SNP

beta 	 regression coefficient of SNP covariate

se 	 standard error of beta

pval 	 p-value of the chi-square statistic

Author(s)

Qiong Yang <qyang@bu.edu> and Ming-Huei Chen <mhchen@bu.edu>

References

kinship package: mixed-effects Cox models, sparse matrices, and modeling data from large pedigrees. Beth Atkinson (atkinson@mayo.edu) for pedigree functions.Terry Therneau (therneau@mayo.edu) for all other functions. 2007. Ref Type: Computer Program http://cran.r-project.org/.

Abecasis, G. R., Cardon, L. R., Cookson, W. O., Sham, P. C., & Cherny, S. S. Association analysis in a variance components framework. Genet Epidemiol, 21 Suppl 1, S341-S346 (2001).

[Package GWAF version 1.2 Index]