lmekin {GWAF}	R Documentation

function for linear mixed effects modeling with a kinship coefficient matrix

Description

A similar function to lmekin from kinship package, but using Wald test, instead of t test. Please see http://cran.r-project.org/web/packages/kinship/kinship.pdf for more information.

See Also

lmekin() function from package kinship

[Package GWAF version 1.0 Index]